// src/components/OrderBookPanel/orderBookPanel.module.scss.d.ts
/** @format */

// autogenerated by typings-for-css-modules-loader.
// Please do not change this file!
declare namespace OrderBookPanelModuleScssNamespace {
  export interface IOrderBookPanelModuleScss {
    body: string
    green: string
    greenBar: string
    head: string
    order0: string
    order1: string
    orderBookPanel: string
    paddingTop: string
    red: string
    redBar: string
    table: string
    tableCell: string
    tableRow: string
  }
}

declare const OrderBookPanelModuleScssModule: OrderBookPanelModuleScssNamespace.IOrderBookPanelModuleScss & {
  /** WARNING: Only available when `css-loader` is used without `style-loader` or `mini-css-extract-plugin` */
  locals: OrderBookPanelModuleScssNamespace.IOrderBookPanelModuleScss
}

export = OrderBookPanelModuleScssModule